Huge commit. Added auth0 authorization through a custom header/jwt system. Moved all of routing execution fxns to chalicelib/app_utils.py, while app.py will now just contain routing/scheduling logic and not be tested. Also made changes to deploy.py and template
